Add files via upload
This game is all about childhood memories! It consists of a 3x3 matrix containing 9 cells. It is a two player game with each player assigned with a marker symbol (X or O). I have implemented this game in Python environment. It allows the user to play against computer as well as other human user. Feel free to provide any suggestions or report about bugs.